Remove SignedMixin in favour of base entity signing by default
While the signing methods are empty at the moment, preparation for ActivityPub, where every object would need to be signed separately (as signature is not in payload header but in the object itself), get rid of separate SignedMixin and just have signature and sign() on the base entity. Diaspora entities provide the necessary methods for signing and validation for DiasporaComment currently.