Don't fail whole send process on failure to sign diaspora public payload
In the case that the root content is from ActivityPub world, sending
a reply for it wont currently work due to the problem of signing
the reply by the author, as there is no target guid to refer to.

We should fix this with various compatibility layers but for now
just raise an error in the log and continue to send to the activitypub
receivers.