Attempt to fix some handle_send issues
It seems sometimes AP deliveries are failing due to
an odd exception where assigning a "to" to the
payload fails with lxml related things. This seems
to indicate that some payload variable is leaking from
the diaspora payload generating to the activitypub code.

Try to ensure this wouldn't be possible to happen and
add some error context for Sentry to better identify
why this would happen, if this commit doesn't fix it.